Change Link and Add New Startpage
As my startpage project grew I used the same repo to redevelop my original start page. The old startpage is in a new repo which is now linked to the Chicago Startpage title. My new startpage that was linked to the old repo title has a new title and link created for it.